+++ src/main/java/com/flyhub/ideamanagementsystem/service/GenderService.java
@@ -1,10 +1,13 @@
 package com.flyhub.ideamanagementsystem.service;
 
 import java.util.List;
+import java.util.Objects;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.flyhub.ideamanagementsystem.DaO.GenderRepository;
 import com.flyhub.ideamanagementsystem.Entity.Gender;
+import com.flyhub.ideamanagementsystem.exception.GenderNotFoundException;
 
 @Service
 public class GenderService {
@@ -12,11 +15,12 @@ public class GenderService {
 	GenderRepository genderRepo;
 	
 	public List<Gender> showAllGenders(){
-		return genderRepo.findAll();
+		List<Gender> genders = genderRepo.findAll();	
+		return genders;
 	}
 	
 	public Gender findGender(Long id){
-		return genderRepo.findById(id).get();
+		return genderRepo.findById(id).orElseThrow(()-> new GenderNotFoundException("Gender with id "+id+" doesn't exist."));
 	}
 	
 	public Gender createGender(Gender gender) {
@@ -24,6 +28,8 @@ public class GenderService {
 	}
 	
 	public void deleteGender(Long id) {
-        genderRepo.deleteById(id);
+		if(Objects.nonNull(findGender(id))) {
+			genderRepo.deleteById(id);
+		}        
     }
 }
